Refactor benefits claims api controllers with more separation between v0 and v1 (#6333)
