+++ app/controllers/concerns/exception_handling.rb
@@ -2,6 +2,7 @@
 
 require 'common/exceptions'
 require 'common/client/errors'
+require 'json_schema/json_api_missing_attribute'
 
 module ExceptionHandling
   extend ActiveSupport::Concern
@@ -9,7 +10,8 @@ module ExceptionHandling
   # In addition to Common::Exceptions::BackendServiceException that have sentry_type :none the following exceptions
   # will also be skipped.
   SKIP_SENTRY_EXCEPTION_TYPES = [
-    Breakers::OutageException
+    Breakers::OutageException,
+    JsonSchema::JsonApiMissingAttribute
   ].freeze
 
   private
@@ -37,7 +39,7 @@ module ExceptionHandling
           Common::Exceptions::Unauthorized.new(detail: exception.detail)
         when ActionController::ParameterMissing
           Common::Exceptions::ParameterMissing.new(exception.param)
-        when Common::Exceptions::BaseError
+        when Common::Exceptions::BaseError, JsonSchema::JsonApiMissingAttribute
           exception
         when Breakers::OutageException
           Common::Exceptions::ServiceOutage.new(exception.outage)
@@ -60,7 +62,12 @@ module ExceptionHandling
   # rubocop:enable Metrics/BlockLength
 
   def render_errors(va_exception)
-    render json: { errors: va_exception.errors }, status: va_exception.status_code
+    case va_exception
+    when JsonSchema::JsonApiMissingAttribute
+      render json: va_exception.to_json_api, status: va_exception.code
+    else
+      render json: { errors: va_exception.errors }, status: va_exception.status_code
+    end
   end
 
   def report_original_exception(exception)
