+++ lib/common/exceptions/service_error.rb
@@ -10,9 +10,12 @@ module Common
       def initialize(options = {})
         @detail = options[:detail]
         @source = options[:source]
+        @errors = options[:errors]
       end
 
       def errors
+        return @errors if @errors.present?
+
         Array(SerializableError.new(i18n_data.merge(detail: @detail, source: @source)))
       end
     end
