+++ modules/claims_api/app/controllers/claims_api/v0/forms/disability_compensation_controller.rb
@@ -1,24 +1,29 @@
 # frozen_string_literal: true
 
-require_dependency 'claims_api/base_disability_compensation_controller'
+require 'evss/disability_compensation_form/service'
+require 'evss/disability_compensation_form/service_exception'
+require 'evss/error_middleware'
+require 'common/exceptions'
 require 'jsonapi/parser'
 
 module ClaimsApi
   module V0
     module Forms
-      class DisabilityCompensationController < BaseDisabilityCompensationController
+      class DisabilityCompensationController < ClaimsApi::V0::Forms::Base
         include ClaimsApi::DocumentValidations
-        include ClaimsApi::EndpointDeprecation
 
         FORM_NUMBER = '526'
+        STATSD_VALIDATION_FAIL_KEY = 'api.claims_api.526.validation_fail'
+        STATSD_VALIDATION_FAIL_TYPE_KEY = 'api.claims_api.526.validation_fail_type'
 
-        skip_before_action(:authenticate)
-        before_action :validate_json_schema, only: %i[submit_form_526 validate_form_526]
-        before_action :validate_documents_content_type, only: %i[upload_supporting_documents upload_form_526]
-        before_action :validate_documents_page_size, only: %i[upload_supporting_documents upload_form_526]
         skip_before_action :validate_json_format, only: %i[upload_supporting_documents]
 
+        # POST to submit disability claim.
+        #
+        # @return [JSON] Record in pending state
         def submit_form_526
+          validate_json_schema
+
           auto_claim = ClaimsApi::AutoEstablishedClaim.create(
             status: ClaimsApi::AutoEstablishedClaim::PENDING,
             auth_headers: auth_headers,
@@ -33,7 +38,7 @@ module ClaimsApi
           end
 
           if auto_claim.errors.present?
-            raise Common::Exceptions::UnprocessableEntity.new(detail: auto_claim.errors.messages.to_s)
+            raise ::Common::Exceptions::UnprocessableEntity.new(detail: auto_claim.errors.messages.to_s)
           end
 
           ClaimsApi::ClaimEstablisher.perform_async(auto_claim.id)
@@ -41,7 +46,14 @@ module ClaimsApi
           render json: auto_claim, serializer: ClaimsApi::AutoEstablishedClaimSerializer
         end
 
+        # PUT to upload a wet-signed 526 form.
+        # Required if first ever claim for Veteran.
+        #
+        # @return [JSON] Claim record
         def upload_form_526
+          validate_documents_content_type
+          validate_documents_page_size
+
           pending_claim = ClaimsApi::AutoEstablishedClaim.pending?(params[:id])
 
           if pending_claim && (pending_claim.form_data['autoCestPDFGenerationDisabled'] == true)
@@ -62,7 +74,13 @@ module ClaimsApi
           render json: unprocessable_response(e), status: :unprocessable_entity
         end
 
+        # POST to upload additional documents to support relevent disability compensation claim.
+        #
+        # @return [JSON] Claim record
         def upload_supporting_documents
+          validate_documents_content_type
+          validate_documents_page_size
+
           claim = ClaimsApi::AutoEstablishedClaim.get_by_id_or_evss_id(params[:id])
           documents.each do |document|
             claim_document = claim.supporting_documents.build
@@ -74,14 +92,103 @@ module ClaimsApi
           render json: claim, serializer: ClaimsApi::ClaimDetailSerializer, uuid: claim.id
         end
 
+        # POST to validate 526 submission payload.
+        #
+        # @return [JSON] Success if valid, error messages if invalid.
+        # rubocop:disable Metrics/MethodLength
         def validate_form_526
           add_deprecation_headers_to_response(response: response, link: ClaimsApi::EndpointDeprecation::V0_DEV_DOCS)
-          super
+          validate_json_schema
+
+          service = EVSS::DisabilityCompensationForm::Service.new(auth_headers)
+          auto_claim = ClaimsApi::AutoEstablishedClaim.new(
+            status: ClaimsApi::AutoEstablishedClaim::PENDING,
+            auth_headers: auth_headers,
+            form_data: form_attributes
+          )
+          service.validate_form526(auto_claim.to_internal)
+          render json: valid_526_response
+        rescue ::EVSS::DisabilityCompensationForm::ServiceException, EVSS::ErrorMiddleware::EVSSError => e
+          error_details = e.is_a?(EVSS::ErrorMiddleware::EVSSError) ? e.details : e.messages
+          track_526_validation_errors(error_details)
+          render json: { errors: format_526_errors(error_details) }, status: :unprocessable_entity
+        rescue ::Common::Exceptions::GatewayTimeout,
+               ::Timeout::Error,
+               ::Faraday::TimeoutError,
+               Breakers::OutageException => e
+          req = { auth: auth_headers, form: form_attributes, source: source_name, auto_claim: auto_claim.as_json }
+          PersonalInformationLog.create(
+            error_class: "validate_form_526 #{e.class.name}", data: { request: req, error: e.try(:as_json) || e }
+          )
+          raise e
+        end
+        # rubocop:enable Metrics/MethodLength
+
+        private
+
+        def flashes
+          initial_flashes = form_attributes.dig('veteran', 'flashes') || []
+          homelessness = form_attributes.dig('veteran', 'homelessness')
+          is_terminally_ill = form_attributes.dig('veteran', 'isTerminallyIll')
+
+          initial_flashes.push('Homeless') if homelessness.present?
+          initial_flashes.push('Terminally Ill') if is_terminally_ill.present? && is_terminally_ill
+
+          initial_flashes.present? ? initial_flashes.uniq : []
         end
 
-        def schema
-          add_deprecation_headers_to_response(response: response, link: ClaimsApi::EndpointDeprecation::V0_DEV_DOCS)
-          super
+        def special_issues_per_disability
+          (form_attributes['disabilities'] || []).map { |disability| special_issues_for_disability(disability) }
+        end
+
+        def special_issues_for_disability(disability)
+          primary_special_issues = disability['specialIssues'] || []
+          secondary_special_issues = []
+          (disability['secondaryDisabilities'] || []).each do |secondary_disability|
+            secondary_special_issues += (secondary_disability['specialIssues'] || [])
+          end
+          special_issues = primary_special_issues + secondary_special_issues
+
+          mapper = ClaimsApi::SpecialIssueMappers::Bgs.new
+          {
+            code: disability['diagnosticCode'],
+            name: disability['name'],
+            special_issues: special_issues.map { |special_issue| mapper.code_from_name!(special_issue) }
+          }
+        end
+
+        def valid_526_response
+          {
+            data: {
+              type: 'claims_api_auto_established_claim_validation',
+              attributes: {
+                status: 'valid'
+              }
+            }
+          }.to_json
+        end
+
+        def format_526_errors(errors)
+          errors.map do |error|
+            { status: 422, detail: "#{error['key']} #{error['detail']}", source: error['key'] }
+          end
+        end
+
+        def track_526_validation_errors(errors)
+          StatsD.increment STATSD_VALIDATION_FAIL_KEY
+
+          errors.each do |error|
+            key = error['key']&.gsub(/\[(.*?)\]/, '')
+            StatsD.increment STATSD_VALIDATION_FAIL_TYPE_KEY, tags: ["key: #{key}"]
+          end
+        end
+
+        def unprocessable_response(e)
+          log_message_to_sentry('Upload error in 526', :error, body: e.message)
+
+          {
+            errors: [{ status: 422, detail: e&.message, source: e&.key }]
+          }.to_json
         end
       end
     end
