+++ modules/claims_api/app/controllers/claims_api/v0/forms/intent_to_file_controller.rb
@@ -1,22 +1,20 @@
 # frozen_string_literal: true
 
-require_dependency 'claims_api/base_form_controller'
 require 'evss/intent_to_file/service'
 
 module ClaimsApi
   module V0
     module Forms
-      class IntentToFileController < ClaimsApi::BaseFormController
-        include ClaimsApi::EndpointDeprecation
-        skip_before_action(:authenticate)
-        before_action :validate_json_format, if: -> { request.post? }
-        before_action :validate_json_schema, only: %i[submit_form_0966 validate]
-        before_action :check_for_type, only: %i[active]
-
+      class IntentToFileController < ClaimsApi::V0::Forms::Base
         FORM_NUMBER = '0966'
         ITF_TYPES = %w[compensation pension burial].freeze
 
+        # POST to submit intent to file claim.
+        #
+        # @return [JSON] Response from BGS
         def submit_form_0966
+          validate_json_schema
+
           bgs_response = bgs_service.intent_to_file.insert_intent_to_file(intent_to_file_options)
           render json: bgs_response,
                  serializer: ClaimsApi::IntentToFileSerializer
@@ -32,28 +30,33 @@ module ClaimsApi
           render json: error, status: :unprocessable_entity
         end
 
+        # GET current intent to file status based on type.
+        #
+        # @return [JSON] Response from BGS
         def active
+          check_for_type
+
           bgs_response = bgs_service.intent_to_file.find_intent_to_file_by_ptcpnt_id_itf_type_cd(
             target_veteran.participant_id,
             ClaimsApi::IntentToFile::ITF_TYPES[active_param]
           )
-          if bgs_response.is_a?(Array)
-            bgs_active = bgs_response.detect do |itf|
-              active?(itf)
-            end
-          elsif active?(bgs_response)
-            bgs_active = bgs_response
-          end
+          bgs_active = if bgs_response.is_a?(Array)
+                         bgs_response.detect { |itf| active?(itf) }
+                       elsif active?(bgs_response)
+                         bgs_response
+                       end
+          message = "No Intent to file is on record for #{target_veteran_name} of type #{active_param}"
+          raise ::Common::Exceptions::RecordNotFound, 'active', detail: message if bgs_active.blank?
 
-          if bgs_active.present?
-            render json: bgs_active, serializer: ClaimsApi::IntentToFileSerializer
-          else
-            render json: itf_not_found, status: :not_found
-          end
+          render json: bgs_active, serializer: ClaimsApi::IntentToFileSerializer
         end
 
+        # POST to validate 0966 submission payload.
+        #
+        # @return [JSON] Success if valid, error messages if invalid.
         def validate
           add_deprecation_headers_to_response(response: response, link: ClaimsApi::EndpointDeprecation::V0_DEV_DOCS)
+          validate_json_schema
           render json: validation_success
         end
 
@@ -64,6 +67,23 @@ module ClaimsApi
 
         private
 
+        def participant_claimant_id
+          return target_veteran.participant_id unless form_type == 'burial'
+
+          raise ::Common::Exceptions::Forbidden, detail: "Representative cannot file for type 'burial'"
+        end
+
+        def intent_to_file_options
+          {
+            intent_to_file_type_code: ClaimsApi::IntentToFile::ITF_TYPES[form_type],
+            participant_claimant_id: form_attributes['participant_claimant_id'] || participant_claimant_id,
+            participant_vet_id: form_attributes['participant_vet_id'] || target_veteran.participant_id,
+            received_date: received_date || Time.zone.now.strftime('%Y-%m-%dT%H:%M:%S%:z'),
+            submitter_application_icn_type_code: ClaimsApi::IntentToFile::SUBMITTER_CODE,
+            ssn: target_veteran.ssn
+          }
+        end
+
         def active?(itf)
           itf.present? && itf[:itf_status_type_cd] == 'Active' && itf[:exprtn_dt].to_datetime > Time.zone.now
         end
@@ -73,17 +93,10 @@ module ClaimsApi
         end
 
         def check_for_type
-          if active_param && !ITF_TYPES.include?(active_param)
-            error = {
-              errors: [
-                {
-                  status: 422,
-                  details: "Must include either compensation, pension or burial as a 'type' parameter."
-                }
-              ]
-            }
-            render json: error, status: :unprocessable_entity
-          end
+          return unless active_param && !ITF_TYPES.include?(active_param)
+
+          message = "Must include either compensation, pension or burial as a 'type' parameter."
+          raise ::Common::Exceptions::UnprocessableEntity.new(detail: message)
         end
 
         def form_type
