+++ modules/claims_api/app/controllers/concerns/claims_api/poa_verification.rb
@@ -7,8 +7,6 @@ module ClaimsApi
     extend ActiveSupport::Concern
 
     included do
-      before_action :verify_power_of_attorney_using_bgs_service, if: :header_request?
-
       def verify_power_of_attorney_using_bgs_service
         logged_in_representative_user = @current_user
         target_veteran_to_be_verified = target_veteran
