+++ modules/claims_api/spec/requests/v1/intent_to_file_request_spec.rb
@@ -69,7 +69,6 @@ RSpec.describe 'Intent to file', type: :request do
     it 'posts a 403 when a representative who doesn\'t have an MPI account tries to post type "burial"' do
       with_okta_user(scopes) do |auth_header|
         VCR.use_cassette('bgs/intent_to_file_web_service/insert_intent_to_file') do
-          expect_any_instance_of(OpenidUser).to receive(:participant_id).and_raise(ArgumentError.new('whatever'))
           data[:data][:attributes] = { type: 'burial' }
           post path, params: data.to_json, headers: headers.merge(auth_header)
           expect(response.status).to eq(403)
